+++ tests/ui/visibility-1.rs
@@ -1,11 +1,25 @@
 use easy_ext::ext;
 
-#[ext(StrExt)]
+#[ext(AssocLevel1)]
 impl str {
-    pub const FOO: &'static str = "_";
-    const Bar: &'static str = "_"; //~ ERROR All items must have a visibility of `pub`
+    pub const ASSOC1: u8 = 1;
+    const ASSOC2: u8 = 2; //~ ERROR all associated items must have a visibility of `pub`
 
-    pub fn baz(&self) -> String {}
+    pub fn assoc(&self) {}
+}
+
+#[ext(AssocLevel2)]
+impl str {
+    fn assoc1(&self) {}
+
+    pub fn assoc2(&self) {} //~ ERROR all associated items must have inherited visibility
+}
+
+#[ext(pub ImplLevel1)]
+impl str {
+    fn assoc1(&self) {}
+
+    pub fn assoc2(&self) {} //~ ERROR all associated items must have inherited visibility
 }
 
 fn main() {}