+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- [Support specifying visibility at impl-level.](https://github.com/taiki-e/easy-ext/pull/25)
+
 ## [0.2.5] - 2021-01-05
 
 - Exclude unneeded files from crates.io.